Game Compiles But Linking Fails

+++ include/typedefs.h
@@ -19,7 +19,7 @@ typedef unsigned short u_short;
 typedef unsigned char u_char;
 typedef unsigned int u_int;
 
-typedef long int dword[2] __attribute__ ((aligned(16)));;
+typedef long long int dword[2] __attribute__ ((aligned(16)));;
 typedef int qword[4] __attribute__ ((aligned(16)));
 typedef int sceVu0IVECTOR[4] __attribute__((aligned (16)));
 typedef float sceVu0FVECTOR[4] __attribute__((aligned (16)));
